requirements.txt
flask
easyocr
groq
python-dotenv
werkzeug
mysql-connector-python
gunicorn
